New Health Bars and minor bug fixes
Created New Branch for UI

fixed scaling for player health bar, it no longer clips inside the players head.

Changed collision for player and enemy to prevent the player from taking damage when walking into objects that are not enemies.

Added health bars to Tree and Stone prefabs as well as adjusted collision for both to better match the sprites.

-Josh Hall